Back-end/requirements.txt
Flask==2.0.2
Flask_Bcrypt==0.7.1
Flask_Cors==3.0.10
Flask_RESTful==0.3.9
flask_uploads==0.2.1
PyJWT==2.3.0
SQLAlchemy==1.4.25
Werkzeug==2.0.2
git+https://github.com/maxcountryman/flask-uploads.git@f66d7dc
